Let SleepHandler to use the symbol counter
Instead of switching timer2 to asynchronous mode on every sleep, we now
configure the symbol counter to always run in asynchronous mode and use
that for wake up timers. It doesn't have a prescaler, but because it is
a 32-bit timer we can still sleep up to 9.5 hours, without a granularity
of 16μs.

Instead of passing the time-to-sleep around, now the target sleep
counter is passed around, which should also make things more accurate.

This breaks uptime.millis.sleep for now, since this counter is no longer
kept. We could calculate this from the total time and millis() counter,
except that they have different overflow ranges, making the subtraction
produce bogus results after the first overflow.

A side effect of this is that the frequency of timer2 drops from 250kHz
to 32kiHz. This affects the blue led only, but the PWM frequency is
still 128Hz which should be enough.